[DDCNL-8706]Updated tests
